Fixed errors and formatting in CollisionDetector and ArtificialIntelligence classes. Restructured gameplayModelTest hierarchy.
